# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * FreshRoute AI accessibility intelligence prototype API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Observation {
  id: number;
  type: string;
  latitude: number;
  longitude: number;
  status: string;
  severity: string;
  source: string;
  initial_confidence: number;
  current_confidence: number;
  created_at: string;
  last_verified_at: string;
  verification_count: number;
  description: string;
  accessibility_impact: string;
  freshness_label: string;
  is_demo: boolean;
}

export interface ObservationInput {
  type: string;
  latitude: number;
  longitude: number;
  status: string;
  severity: string;
  source: string;
  initial_confidence: number;
  description: string;
  accessibility_impact: string;
}

export interface Report {
  id: number;
  report_type: string;
  latitude: number;
  longitude: number;
  severity: string;
  description: string;
  created_at: string;
  status: string;
  is_demo: boolean;
}

export interface ReportInput {
  report_type: string;
  latitude: number;
  longitude: number;
  severity: string;
  description: string;
}

export interface MapPath {
  id: string;
  status: string;
  confidence: number;
  /**
     * @items.minItems 2
     * @items.maxItems 2
     */
  coordinates: number[][];
}

export interface MapData {
  center: number[];
  zoom: number;
  observations: Observation[];
  paths: MapPath[];
}

export interface DashboardStats {
  accessibility_observations: number;
  active_obstacles: number;
  verified_accessible_paths: number;
  average_confidence: number;
  recently_changed_locations: number;
  last_updated: string;
}

export interface AnalyticsPoint {
  label: string;
  value: number;
}

export interface Analytics {
  observations_over_time: AnalyticsPoint[];
  obstacles_over_time: AnalyticsPoint[];
  confidence_over_time: AnalyticsPoint[];
  reports_by_category: AnalyticsPoint[];
  status_distribution: AnalyticsPoint[];
}

export interface RouteRequest {
  start: string;
  destination: string;
  mobility_requirement: string;
  preference: string;
}

export interface RouteAlternative {
  id: string;
  name: string;
  distance_km: number;
  duration_minutes: number;
  accessibility_score: number;
  confidence: number;
  status: string;
  reason: string;
  /**
     * @items.minItems 2
     * @items.maxItems 2
     */
  coordinates: number[][];
}

export interface RouteResult {
  recommended_route_id: string;
  alternatives: RouteAlternative[];
  explanation: string;
}

export interface DemoResult {
  stage: string;
  timeline: string[];
  changed_observation: Observation;
  route_result: RouteResult;
  message: string;
}

export interface VisionInput {
  file_name: string;
  image_hint?: string;
}

export interface VisionDetection {
  object: string;
  confidence: number;
  impact: string;
  note: string;
}

export interface VisionResult {
  mode: string;
  detections: VisionDetection[];
  message: string;
}

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;