# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * FreshRoute AI accessibility intelligence prototype API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsPoint } from './analyticsPoint';

export interface Analytics {
  observations_over_time: AnalyticsPoint[];
  obstacles_over_time: AnalyticsPoint[];
  confidence_over_time: AnalyticsPoint[];
  reports_by_category: AnalyticsPoint[];
  status_distribution: AnalyticsPoint[];
}