# lib/api-zod/src/generated/types/analyticsPoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * FreshRoute AI accessibility intelligence prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsPoint {
  label: string;
  value: number;
}