# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * FreshRoute AI accessibility intelligence prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  accessibility_observations: number;
  active_obstacles: number;
  verified_accessible_paths: number;
  average_confidence: number;
  recently_changed_locations: number;
  last_updated: string;
}